# swagger/center.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OKR 系统API",
    "description": "OKR 系统相关接口文档！",
    "contact": {
      "name": "发际线与我作对",
      "url": "https://github.com/CarefreeState/OKR-System",
      "email": "2040484356@qq.com"
    },
    "version": "3.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1701",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/core/celebrate/{day}": {
      "post": {
        "tags": [
          "OKR 内核"
        ],
        "summary": "确定庆祝日",
        "operationId": "confirmCelebrateDayUsingPOST",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "description": "庆祝日（星期）",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OKR 内核带场景值的数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/core/complete": {
      "post": {
        "tags": [
          "OKR 内核"
        ],
        "summary": "结束 OKR",
        "operationId": "completeUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OKR 内核带场景值的数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/core/create": {
      "get": {
        "tags": [
          "OKR 内核"
        ],
        "summary": "创建一个OKR",
        "operationId": "createOkrUsingGET",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OKR 操作数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/core/search": {
      "get": {
        "tags": [
          "OKR 内核"
        ],
        "summary": "查看一个OKR内核",
        "operationId": "searchOkrCoreUsingGET",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OKR 内核带场景值的数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«OKR 内核详细数据»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/core/summary": {
      "post": {
        "tags": [
          "OKR 内核"
        ],
        "summary": "总结 OKR",
        "operationId": "summaryOKRUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/总结 OKR 所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/task/{option}/add": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "增加一条任务",
        "operationId": "addTaskUsingPOST",
        "parameters": [
          {
            "name": "option",
            "in": "path",
            "description": "任务选项（0:action, 1:P1, 2:P2）",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/增加任务所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/task/{option}/remove": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "删除一个任务",
        "operationId": "removeTaskUsingPOST",
        "parameters": [
          {
            "name": "option",
            "in": "path",
            "description": "任务选项（0:action, 1:P1, 2:P2）",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/删除任务所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/task/{option}/update": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "更新一条任务",
        "operationId": "updateTaskUsingPOST",
        "parameters": [
          {
            "name": "option",
            "in": "path",
            "description": "任务选项（0:Action, 1:P1, 2:P2）",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/更新任务所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "任务数据": {
        "title": "任务数据",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "任务内容"
          },
          "quadrantId": {
            "type": "integer",
            "description": "象限 ID",
            "format": "int64"
          }
        }
      },
      "任务更新数据": {
        "title": "任务更新数据",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "任务内容"
          },
          "id": {
            "type": "integer",
            "description": "任务 ID",
            "format": "int64"
          },
          "isCompleted": {
            "type": "boolean",
            "description": "是否完成",
            "example": false
          }
        }
      },
...
      "删除任务所需数据": {
        "title": "删除任务所需数据",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "任务 ID",
            "format": "int64"
          },
          "scene": {
            "type": "string",
            "description": "场景"
          }
        }
      },
...
      "增加任务所需数据": {
        "title": "增加任务所需数据",
        "type": "object",
        "properties": {
          "scene": {
            "type": "string",
            "description": "场景"
          },
          "taskDTO": {
            "description": "任务数据",
            "$ref": "#/components/schemas/任务数据"
          }
        }
      },
...
      "更新任务所需数据": {
        "title": "更新任务所需数据",
        "type": "object",
        "properties": {
          "scene": {
            "type": "string"
          },
          "taskUpdateDTO": {
            "description": "更新任务的数据",
            "$ref": "#/components/schemas/任务更新数据"
          }
        }
      },
...
      "统一响应": {
        "title": "统一响应",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "具体数据"
          },
          "message": {
            "type": "string",
            "description": "描述语"
          }
        }
      },
...
    }
  }
}